requirements.txt
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pillow
typer
typer-cli
pyyaml